LayoutTests/http/tests/security/contentSecurityPolicy/blob-urls-match-self-expected.txt

ALERT: PASS (1/1)!
blob: URLs are same-origin with the page in which they were created, and should therefore match the 'self' source in CSP directives.
